### 2025-03-11 5530b22
pare down paper text a bit
