+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`spike`: A tool to drizzle _HST_, _JWST_, and Roman point spread functions for improved analyses'
+title: '`spike`: A tool to drizzle _HST_, _JWST_, and Roman PSFs for improved analyses'
 tags:
   - Python
   - astronomy
@@ -48,21 +48,23 @@ The built-in PSF generation options are `TinyTim` [@Krist.etal.2011] and the @Gi
 
 In most cases, users will only ever interact with the top-level functions `spike.psf.hst`, `spike.psf.jwst`, and `spike.psf.roman`. However, PSF model creation can be accessed directly via the functions in `spike.psfgen`, and this feature may be of added value to users on its own, including for use with telescopes/instruments not explicitly mentioned here, as `spike` smoothes over some of the complication of individual tools as an all-in-one means of accessing model PSFs via simple Python functions. One can, in principle, specify only minimal information (data directory, object name/coordinates, instrument/camera, and PSF generation method) and return a drizzled PSF from the default remaining default settings/arguments.
 
-`spike` is also parallelized for use with large sets of coordinates using `multiprocess`, and the option to run the code in parallel is togglable. 
+<!-- `spike` is also parallelized for use with large sets of coordinates using `multiprocess`, and the option to run the code in parallel is togglable.  -->
 
-Optionally, `spike` can complete the "drizzle" steps on the original imaging data, as well, making for one-time, _consistent_ processing of both the PSFs and the original images. As some users will prefer to analyze high-level data products directly from the _HST_, _JWST_, and, eventually, Roman pipelines, this feature is opt-in.
+<!-- Optionally, `spike` can complete the "drizzle" steps on the original imaging data, as well, making for one-time, _consistent_ processing of both the PSFs and the original images. As some users will prefer to analyze high-level data products directly from the _HST_, _JWST_, and, eventually, Roman pipelines, this feature is opt-in. -->
 
-In addition to .fits outputs, users can specify whether they would like a .asdf file returned, too, and it is possible to write .png images of the individual PSF models and the final co-add, which allows users to do quick visual inspection.
+<!-- In addition to .fits outputs, users can specify whether they would like a .asdf file returned, too, and it is possible to write .png images of the individual PSF models and the final co-add, which allows users to do quick visual inspection. -->
 
 
 # Preparation for Future Observatories
 
-Right now, the Roman pipeline uses the same resampling/drizzling implementation as _JWST_ via `romancal` [@romancal], and `WebbPSF` includes a module to simulate Roman PSFs. As a forward looking step, a module is included here to handle the proper co-addition of Roman PSFs. The `spike.psf.roman` module includes step for subpixel alignment, as has been done for _HST_ and _JWST_ by "tweaking" images, and model Roman PSFs are then directly drizzled using a version of `romancal.resample`. When the observatory is actually launched and data become available, `spike` will be updated to reflect the most current version of the Nancy Grace Roman Space Telescope pipeline and the detailed considerations of the real data. Due to the anticipated single pointing file size, `spike.psf.roman` takes as input single detector images, which can be used with `spike.psfgen` -- the working directory need not be local; it just needs to have read/write access. Since Roman is not yet collecting data, `spike.psf.roman` is based on the structure of simulated single detector data from [e.g., @Troxel.etal.2023].
[…]
