### 2025-03-11 ec2596b
remove commented text from paper draft -- first pass at removing text
