### 2025-03-12 7152e8e
add try/except for the STDPSF fix in case photutils accepts my PR or it changes in a new version
