### 2025-03-14 63426b0
update readthedocs yaml to fix compilation
