### 2025-03-14 59c6efc
update stdatamodel version to make sure that modules don't change
