### 2025-03-14 cab25ca
set fixed version for stdatamodel dependecy
