### 2025-03-14 4f3fe5f
add specific versions to all requirements to ensure docs
