+++ docs/requirements.txt
@@ -1,20 +1,142 @@
-sphinx
-sphinx-rtd-theme
-asdf
-astropy
-crds
-drizzle
-drizzlepac @ git+https://github.com/spacetelescope/drizzlepac.git
-gwcs
-jsonschema
-matplotlib
-numpy
-photutils
-psutil
-pyyaml
-roman-datamodels
-scipy
-spherical-geometry
-stdatamodels==2.2.0 
-tweakwcs
-webbpsf
+alabaster==1.0.0
+asdf==3.5.0
+asdf-astropy==0.7.0
+asdf_coordinates_schemas==0.3.0
+asdf_standard==1.1.1
+asdf_transform_schemas==0.5.0
+asdf_wcs_schemas==0.4.0
+astrocut==0.9
+astropy==6.1.7
+astropy-iers-data==0.2024.12.30.0.33.36
+astroquery==0.4.7
+asttokens==3.0.0
+attrs==24.3.0
+babel==2.16.0
+backports.tarfile==1.2.0
+beautifulsoup4==4.12.3
+bokeh==3.6.2
+build==1.2.2.post1
+certifi==2024.12.14
+charset-normalizer==3.4.1
+contourpy==1.3.1
+crds==12.0.9
+cycler==0.12.1
+decorator==5.1.1
+docutils==0.21.2
+drizzle==2.0.0
+drizzlepac @ git+https://github.com/spacetelescope/drizzlepac.git@7951b66520e4c245933f6e14d6990e32ac9f5779
+et_xmlfile==2.0.0
+exceptiongroup==1.2.2
+executing==2.1.0
+filelock==3.16.1
+fitsblender==0.4.4
+fonttools==4.55.3
+gwcs==0.22.1
+html5lib==1.1
+id==1.5.0
+idna==3.10
+imageio==2.36.1
+imagesize==1.4.1
+importlib_metadata==8.5.0
+iniconfig==2.0.0
+ipython==8.31.0
+jaraco.classes==3.4.0
+jaraco.context==6.0.1
+jaraco.functools==4.1.0
+jedi==0.19.2
+Jinja2==3.1.5
+jmespath==1.0.1
+joblib==1.4.2
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+keyring==25.6.0
+kiwisolver==1.4.8
+lazy_loader==0.4
+lxml==5.3.0
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+matplotlib==3.10.0
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+more-itertools==10.5.0
+networkx==3.4.2
+nh3==0.2.21
+numpy==2.2.1
+openpyxl==3.1.5
+packaging==24.2
+pandas==2.2.3
+Parsley==1.3
+parso==0.8.4
+pexpect==4.9.0
+photutils==2.0.2
+pillow==11.0.0
+pluggy==1.5.0
+poppy==1.1.1
+prompt_toolkit==3.0.48
+psutil==6.1.1
+ptyprocess==0.7.0
+pure_eval==0.2.3
+pyerfa==2.0.1.5
+Pygments==2.18.0
+pyparsing==3.2.0
+PyPDF2==3.0.1
+pyproject_hooks==1.2.0
+pysiaf==0.24.1
+pytest==8.3.5
+python-dateutil==2.9.0.post0
+pytz==2024.2
+pyvo==1.6
+PyYAML==6.0.2
+rad==0.22.0
+readme_renderer==44.0
+referencing==0.35.1
+requests==2.32.3
+requests-toolbelt==1.0.0
+rfc3986==2.0.0
+rich==13.9.4
[…]
