+++ docs/requirements_generic.txt
@@ -1,0 +1,20 @@
+sphinx
+sphinx-rtd-theme
+asdf
+astropy
+crds
+drizzle
+drizzlepac @ git+https://github.com/spacetelescope/drizzlepac.git
+gwcs
+jsonschema
+matplotlib
+numpy
+photutils
+psutil
+pyyaml
+roman-datamodels
+scipy
+spherical-geometry
+stdatamodels==2.2.0 
+tweakwcs
+webbpsf
