### 2025-03-18 91060fa
update to 1.0.2, which has PyPI description
