+++ setup.py
@@ -1,9 +1,11 @@
 import setuptools
-
+from pathlib import Path
+spike_home = Path(__file__).parent
+pypi_descrip = (spike_home / "README.md").read_text()
 
 setuptools.setup(
 	name = "spike-psf",
-	version = "1.0",
+	version = "1.0.2",
 	author = "Ava Polzin",
 	author_email = "apolzin@uchicago.edu",
 	description = "Drizzle/resample HST, JWST, and Roman PSFs for improved analyses.",
@@ -21,8 +23,11 @@ setuptools.setup(
 		"Topic :: Scientific/Engineering :: Physics"],
 	python_requires = ">=3.10",
 	install_requires = ["asdf", "astropy", "crds", "drizzle", 
-	"drizzlepac @ git+https://github.com/spacetelescope/drizzlepac.git", "gwcs", 
+	"drizzlepac @ git+https://github.com/spacetelescope/drizzlepac.git", 
+	"gwcs", 
 	"jsonschema", "matplotlib", "numpy", "photutils", "psutil", "pyyaml", "roman-datamodels", 
 	"scipy", "spherical-geometry", "stdatamodels==2.2.0", "tweakwcs", "webbpsf"],
-	package_data={'spike': ['configs/*/*']}
+	package_data={'spike': ['configs/*/*']},
+	long_description=pypi_descrip,
+    long_description_content_type='text/markdown'
 )
