### 2025-03-22 5baf885
update discussion of issues to request details
