### 2025-03-22 9d49b62
add explicit error message for drc/drz files as inputs
