### 2025-03-28 2443443
remove hacky photutils solution since STDPSFs have been renamed on STScI site
