+++ setup.py
@@ -5,7 +5,7 @@ pypi_descrip = (spike_home / "README.md").read_text()
 
 setuptools.setup(
 	name = "spike-psf",
-	version = "1.0.3",
+	version = "1.0.4",
 	author = "Ava Polzin",
 	author_email = "apolzin@uchicago.edu",
 	description = "Drizzle/resample HST, JWST, and Roman PSFs for improved analyses.",
