### 2025-03-30 17f0f40
move footnotes in paper draft
