### 2025-03-30 e769d1a
clarify how cutout option works in docs and docstrings
