### 2025-03-31 437f03d
add explicit extension number to rewrite_fits in response to #5
