### 2025-03-31 19166c8
revert change to extension version as it apparently does not resolve issue
