### 2025-04-02 c07d967
fix 'bug' that resulted in declination added to filename twice
