+++ docs/index.rst
@@ -36,7 +36,7 @@ If you install all of the optional dependencies described above, your shell's st
 	# export CRDS_SERVER_URL="https://jwst-crds.stsci.edu"
 	# export CRDS_SERVER_URL="https://roman-crds.stsci.edu"
 
-Since both JWST and Roman CRDS servers may be defined, these variables are added directly within ``spike.jwstcal`` and ``spike.romancal`` and do not need to be added to your startup file.
+Since both JWST and Roman CRDS servers may be defined, these variables are added directly within ``spike.psf.jwst`` and ``spike.psf.roman`` and do not need to be added to your startup file.
 
 .. toctree::
 	:maxdepth: 1
