### 2025-04-16 831f8d0
incrementally move toward JOSS word limit by cutting some of the paper
