+++ paper/paper.md
@@ -35,7 +35,7 @@ For analyses to be consistent, mock PSFs may be used, but the proper treatment o
 
 The premise of `spike` is that, given a directory containing reduced and calibrated, but not yet co-added, .fits files from _HST_, _JWST_, or Roman and the coordinates of an object of interest, detector/chip-specifc model PSFs themselves can be directly "drizzled". The code can be run using images that have been "tweaked" or images that have not yet undergone any resampling or post-processing.
 
-The package relies on `astropy` [@AstropyCollaboration.2013; @AstropyCollaboration.2018; @AstropyCollaboration.2022] for conversion of objects' astronomical coordinates (in right ascension and declination) to pixel coordinates (in X, Y), using the images' world coordinate system information. A model PSF is then generated for each unique input image and coordinates combination. Instrument information, including camera, filter, and, if necessary, chip, is automatically read directly from the header of each .fits file by default, but can be overridden by user choice. Similarly, users can generate PSF models using pre-defined defaults, update model parameters, or upload their own model PSFs to be drizzled.
+A model PSF is generated for each unique input image and coordinate combination. Instrument information, including camera, filter, and, if necessary, chip, is automatically read directly from the header of each .fits file by default, but can be overridden by user choice. Similarly, users can generate PSF models using pre-defined defaults, update model parameters, or upload their own model PSFs to be drizzled.
 
 Though a generic empirical PSF can be computed from the drizzled image, `spike` includes the ability for users to generate and drizzle empirical PSFs via a number of different codes. Computation of these PSFs may require more user input (such as selecting good stars for which to measure light profiles) and may take more time. The empirical models included here are chosen for relevance to high-resolution space-based data; user-generated PSFs from other tools may be used with `spike.psf`, but are not shipped as part of `spike.psfgen`.
 
@@ -47,7 +47,7 @@ The built-in PSF generation options are `TinyTim` [@Krist.etal.2011] and the @Gi
 
 ![Same as \autoref{fig:hstcomparison} for \textit{JWST}/NIRCam imaging in F115W. Note that the ePSF model shown here was generated using a lower detection threshold and a different star selection algorithm due to a paucity of stars in this field. \label{fig:jwstcomparison}](spike_psf_jwstcompare.png)
 
-In most cases, users will only ever interact with the top-level functions `spike.psf.hst`, `spike.psf.jwst`, and `spike.psf.roman`. However, PSF model creation can be accessed directly via the functions in `spike.psfgen`, and this feature may be of added value to users on its own, including for use with telescopes/instruments not explicitly mentioned here, as `spike` smoothes over some of the complication of individual tools as an all-in-one means of accessing model PSFs via simple Python functions. One can, in principle, specify only minimal information (data directory, object name/coordinates, instrument/camera, and PSF generation method) and return a drizzled PSF from the default remaining default settings/arguments.
+In most cases, users will only ever interact with the top-level functions `spike.psf.hst`, `spike.psf.jwst`, and `spike.psf.roman`. However, PSF model creation can be accessed directly via the functions in `spike.psfgen`, and this feature may be of added value to users on its own, including for use with telescopes/instruments not explicitly mentioned here, as `spike` smoothes over some of the complication of individual tools as an all-in-one means of accessing model PSFs via simple Python functions. One can, in principle, specify only minimal information (data directory, object name/coordinates, instrument/camera, and PSF generation method) and return a drizzled PSF from the default settings/arguments.
 
 
 # Preparation for Future Observatories
@@ -57,7 +57,7 @@ When the Nancy Grace Roman Space Telescope is launched and data become available
 
 # Managing Restrictive Dependencies
 
[…]
