### 2025-04-16 2e2c534
further pare down paper text
