### 2025-04-16 #7889 @avapolzin
Hi, @warrickball -- I know you're not officially the editor for this submission, but I figured you might have an answer anyway. As you've suggested, I have gotten the paper below 1500 words, but could potentially save another >200 words, if I remove the last two sections as most of their content is duplicated in the documentation. I wanted to check before I did this, though, as I'm sure JOSS has a desired balance between manuscript length and detail, and if the 1000 word limit is not stringent, it may be preferred that I retain those sections. (I can also take this up with the editor once they're assigned, but wanted to act on what I can now to facilitate the review later.)
