### 2025-04-18 94d1ea2
add contributing guidelines, which reiterate README
