### 2025-05-08 636e9e3
bug fixes and preliminary example notebooks
