+++ paper/paper.md
@@ -26,7 +26,7 @@ Point spread functions (PSFs) describe the distribution of light for a pure poin
 
 The PSF of co-added images is of generic interest to both ground- and space-based instruments, as it can be impacted by differing co-addition schemes and may have an effect on the analysis of those data [@Mandelbaum.etal.2023]. The cumulative effect of the geometric distortions and offsets in angle and pixel location of space-based data are apparent in the effective PSF of the co-added and resampled [drizzled, @Fruchter.Hook.2002] image, making a PSF modeled on uncombined images insufficient for careful photometric analyses. This is a recognized limitation of existing PSF models, and `DrizzlePac` recently added functionality to use drizzled pre-computed model PSFs in their native photometric catalog generator [@DrizzlePac3.3.0]. This drizzled PSF uses generic `TinyTim` [@Krist.etal.2011] models, which do not account for source position on the chip and do not allow the user to set model parameters without overwriting the existing grid of model PSFs. Simultaneously, the resultant co-added PSF is not output, and there is not an existing, simple way to generate drizzled PSFs for use in other analyses. Instead, an empirical PSF may be taken from the data or unsaturated stars near the object of interest may be selected as a proxy for the PSF. For particularly crowded fields, or those where most stars are saturated, these alternatives pose a real problem.
 
-For analyses to be consistent, mock PSFs may be used, but the proper treatment of them involves generating model PSFs and then co-adding and processing them in the same way as the parent images are co-added and processed. Workflows have been created for some instruments, [^1] and individual packages have tried to address this in their own data handling, with e.g.,  `Grizli` [@Brammer_grizli] returning drizzled standard empirical PSFs, but, similar to the drizzled PSFs computed by `DrizzlePac`, such solutions leave little room for users to define the method or specifics of model PSF generation. The code presented here, `spike`, streamlines the potentially onerous process of "realistic" PSF generation, taking images and coordinates and directly outputting correctly co-added model PSFs for the _Hubble Space Telescope_ (_HST_), the _James Webb Space Telescope_ (_JWST_), and the upcoming Nancy Grace Roman Space Telescope. Model PSFs can be generated using different industry-standard packages, empirical PSFs can be computed from input images, or users can provide PSFs associated with the individual images. `spike` [^2] is both easy to use and flexible. 
+For analyses to be consistent, mock PSFs may be used, but the proper treatment of them involves generating model PSFs and then co-adding and processing them in the same way as the parent images are co-added and processed. Workflows have been created for some instruments[^1] and surveys [e.g., @Ji.etal.2024], and individual packages have tried to address this in their own data handling, with e.g.,  `Grizli` [@Brammer_grizli] returning drizzled standard empirical PSFs, but, similar to the drizzled PSFs computed by `DrizzlePac`, such solutions leave little room for users to define the method or specifics of model PSF generation. The code presented here, `spike`, streamlines the potentially onerous process of "realistic" PSF generation, taking images and coordinates and directly outputting correctly co-added model PSFs for the _Hubble Space Telescope_ (_HST_), the _James Webb Space Telescope_ (_JWST_), and the upcoming Nancy Grace Roman Space Telescope. Model PSFs can be generated using different industry-standard packages, empirical PSFs can be computed from input images, or users can provide PSFs associated with the individual images. `spike` [^2] is both easy to use and flexible. 
 
 [^1]: <https://github.com/spacetelescope/hst_notebooks/tree/main/notebooks/WFC3/point_spread_function>; <https://github.com/spacetelescope/wfc3_photometry>
 [^2]: <https://github.com/avapolzin/spike>
