+++ paper/paper.bib
@@ -855,3 +855,23 @@ title = {{grizli}},
 url = {https://github.com/gbrammer/grizli},
 year = {2024}
 }
+
+@ARTICLE{Ji.etal.2024,
+       author = {{Ji}, Zhiyuan and {Williams}, Christina C. and {Tacchella}, Sandro and {Suess}, Katherine A. and {Baker}, William M. and {Alberts}, Stacey and {Bunker}, Andrew J. and {Johnson}, Benjamin D. and {Robertson}, Brant and {Sun}, Fengwu and {Eisenstein}, Daniel J. and {Rieke}, Marcia and {Maseda}, Michael V. and {Hainline}, Kevin and {Hausen}, Ryan and {Rieke}, George and {Willmer}, Christopher N.~A. and {Egami}, Eiichi and {Shivaei}, Irene and {Carniani}, Stefano and {Charlot}, Stephane and {Chevallard}, Jacopo and {Curtis-Lake}, Emma and {Looser}, Tobias J. and {Maiolino}, Roberto and {Willott}, Chris and {Chen}, Zuyi and {Helton}, Jakob M. and {Lyu}, Jianwei and {Nelson}, Erica and {Bhatawdekar}, Rachana and {Boyett}, Kristan and {Sandles}, Lester},
+        title = "{JADES + JEMS: A Detailed Look at the Buildup of Central Stellar Cores and Suppression of Star Formation in Galaxies at Redshifts 3 < z < 4.5}",
+      journal = {\apj},
+     keywords = {Galaxy formation, Galaxy evolution, Galaxy quenching, Galaxy structure, High-redshift galaxies, 595, 594, 2040, 622, 734, Astrophysics - Astrophysics of Galaxies},
+         year = 2024,
+        month = oct,
+       volume = {974},
+       number = {1},
+          eid = {135},
+        pages = {135},
+          doi = {10.3847/1538-4357/ad6e7f},
+archivePrefix = {arXiv},
+       eprint = {2305.18518},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2024ApJ...974..135J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
