### 2025-05-08 460db68
move installation, create kernel docs
