### 2025-05-09 c7b47ca
add keepfinal functionality because of storage constraints
