### 2025-05-09 9ffc758
add example notebook to documentation
