+++ docs/index.rst
@@ -8,8 +8,9 @@ Easily generated and resampled space-based PSFs.
     :target: https://arxiv.org/abs/2503.02288
 .. image:: https://joss.theoj.org/papers/744ad03a43040debb962391d1668ea5c/status.svg
     :target: https://joss.theoj.org/papers/744ad03a43040debb962391d1668ea5c
-.. image:: https://img.shields.io/badge/GitHub-avapolzin/spike-08872b
+.. image:: https://img.shields.io/badge/GitHub-avapolzin/spike-74349c
     :target: https://github.com/avapolzin/spike
+    ..08872b as alternative color
 
 To install ``spike``, either ``pip install spike-psf`` or install the development version from the distribution on `GitHub <https://github.com/avapolzin/spike>`_. More detailed instructions are available on `this page <https://spike-psf.readthedocs.io/en/latest/install.html>`_.
 
@@ -27,10 +28,10 @@ To install ``spike``, either ``pip install spike-psf`` or install the developmen
 	citation.rst
 
 .. .. toctree::
-.. 	:maxdepth: 1
-.. 	:caption: Examples:
+	:maxdepth: 1
+	:caption: Examples:
 
-.. 	acs_example.nblink
+	acs_example.nblink
 .. 	nircam_example.nblink
 .. 	psf_generation.nblink
 
