### 2025-05-09 765d85c
update conf.py for nbsphinx
