### 2025-05-09 8ec2ac9
actually add notebook to documentation
