+++ spike/psfgen/psfgen.py
@@ -692,6 +692,15 @@ def effpsf(coords, img, imcam, pos, plot = False, verbose = False, mask = True, 
 
 	dat = fits.open(img)[ext].data
 
+	if mask:
+		maskarr = fits.open(img)[('DQ', extv)].data
+		dat[maskarr > 0] = np.nan # only retain good pixels
+		maskarr[maskarr > 0] = True
+
+		if usermask:
+			dat[usermask] = np.nan
+			maskarr[usermask] = True
+
 	mean, median, std = sigma_clipped_stats(dat, sigma=3.0)
 
 	if starselect.upper() == 'DAO':
@@ -701,20 +710,13 @@ def effpsf(coords, img, imcam, pos, plot = False, verbose = False, mask = True, 
 	if starselect.upper() == 'IRAF':
 		find = IRAFStarFinder(threshold = thresh*std,  **starselectargs)
 
-	if mask:
-		maskarr = fits.open(img)[('DQ', extv)].data
-		dat[maskarr > 0] = 0 # only retain good pixels
-		maskarr[maskarr > 0] = True
-		if usermask:
-			dat[usermask] = 0
-			maskarr[usermask] = True
-		if verbose:
+	if verbose:
 			print('Identifying stars to use in ePSF')
+			
+	if mask:
 		sources = find(dat, mask = maskarr)
 
 	if not mask:
-		if verbose:
-			print('Identifying stars to use in ePSF')
 		sources = find(dat)
 
 	exsize = int(2 * ((fov_arcsec/plate_scale[pixkey])//2) + 1) #size of extraction box
