### 2025-05-09 b4edf8c
make epsf function somewhat easier
