### 2025-05-10 7738718
update example notebook with output images
