### 2025-05-11 013a938
add nircam example notebook
