+++ docs/index.rst
@@ -31,7 +31,7 @@ To install ``spike``, either ``pip install spike-psf`` or install the developmen
 	:caption: Examples:
 
 	acs_example.nblink
-.. 	nircam_example.nblink
+	nircam_example.nblink
 .. 	psf_generation.nblink
 
 
