### 2025-05-11 498b964
update conf.py with ipython for syntax highlighting
