+++ docs/requirements.txt
@@ -2,6 +2,7 @@ sphinx
 sphinx-rtd-theme
 nbsphinx
 nbsphinx-link
+ipython
 asdf
 astropy
 crds
