### 2025-05-11 7f4f7a8
update example_notebooks READMEto highlight consistency of syntax across entire package
