+++ example_notebooks/README.md
@@ -1,3 +1,5 @@
+You may notice that the example notebooks in this directory are all fairly similar -- this is a function of `spike` using common syntax across modules for both PSF drizzling and PSF generation, which makes the code very simple to use. 
+
 If you have created a new environment for `spike`, you will need to generate a corresponding kernel to run .ipynb files with that environment. If you are using conda, this will look like the following:
 
 ```bash
