### 2025-05-11 b5b4f6f
fix typo in faq
