+++ paper/paper.bib
@@ -875,3 +875,10 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@software{Perrin_STPSF,
+author = {Perrin, Marshall and Melendez, Marcio and Osborne, Shannon and Geda, Robel and Sappington, Bradley and Lajoie, Charles-Philippe and Long, Joseph and Otor, O. Justin},
+license = {BSD-3-Clause},
+title = {{STPSF}},
+url = {https://github.com/spacetelescope/stpsf},
+version = {2.0.0}
+}
