+++ docs/requirements.txt
@@ -21,3 +21,4 @@ spherical-geometry
 stdatamodels==2.2.0 
 tweakwcs
 webbpsf
+stpsf
