+++ docs/requirements_version.txt
@@ -137,6 +137,7 @@ tzdata==2024.2
 urllib3==2.3.0
 wcwidth==0.2.13
 webbpsf==1.5.0
+stpsf==2.0.0
 webencodings==0.5.1
 xyzservices==2024.9.0
 zipp==3.21.0

[…]
