### 2025-05-12 fd4c3f6
add single frame psf generation tutorial to docs
