### 2025-05-13 #7889 @avapolzin
Thank you, @ivastar! Here are some potential reviewers based on the JOSS reviewers list and apparent research history/interests:
- Karl D. Gordon (@\karllark)
- Tom J. Wilson (@\Onoddil)
- Yoonsoo P. Bach (@\ysBach)

I can take another pass, too, if a longer list is preferred.
