### 2025-05-13 a13adaf
actual last bit of clarification in docs - adding links to examples in top-level funcs
