+++ docs/citation.rst
@@ -12,7 +12,7 @@ If you use this package or the scripts in this repository in a publication, plea
 * if you use ``spike`` with Roman data: ``romancal`` (`Roman calibration pipeline developers 2024 <https://github.com/spacetelescope/romancal>`_)
 * if you use ``TinyTim`` model PSFs: ``TinyTim`` (`Krist et al. 2011 <https://ui.adsabs.harvard.edu/abs/2011SPIE.8127E..0JK/abstract>`_)
 * if you use ``TinyTim`` model PSFs with the Gillis et al. modification: ``TinyTim`` (`Krist et al. 2011 <https://ui.adsabs.harvard.edu/abs/2011SPIE.8127E..0JK/abstract>`_) + `Gillis (2019) <https://bitbucket.org/brgillis/tinytim_psfs/src/master/>`_/`Gillis et al. (2020) <https://ui.adsabs.harvard.edu/abs/2020MNRAS.496.5017G/abstract>`_
-* if you use ``STPSF`` (formerly ``WebbPSF``) model PSFs: ``STPSF`` `Perrin et al (2025) <https://github.com/spacetelescope/stpsf>`_, ``WebbPSF`` (Perrin et al. `2012 <https://ui.adsabs.harvard.edu/abs/2012SPIE.8442E..3DP/abstract>`_; `2014 <https://ui.adsabs.harvard.edu/abs/2014SPIE.9143E..3XP/abstract>`_; `2024 <https://github.com/spacetelescope/webbpsf>`_)
+* if you use ``STPSF`` (formerly ``WebbPSF``) model PSFs: ``STPSF`` `(Perrin et al. 2025) <https://github.com/spacetelescope/stpsf>`_, ``WebbPSF`` (Perrin et al. `2012 <https://ui.adsabs.harvard.edu/abs/2012SPIE.8442E..3DP/abstract>`_; `2014 <https://ui.adsabs.harvard.edu/abs/2014SPIE.9143E..3XP/abstract>`_; `2024 <https://github.com/spacetelescope/webbpsf>`_)
 * if you use ``PSFEx`` to generate empirical PSFs: ``PSFEx`` (Bertin `2011 <https://ui.adsabs.harvard.edu/abs/2011ASPC..442..435B/abstract>`_; `2013 <https://ui.adsabs.harvard.edu/abs/2013ascl.soft01001B/abstract>`_) and ``SExtractor`` (`Bertin & Arnauts 1996 <https://ui.adsabs.harvard.edu/abs/1996A%26AS..117..393B/abstract>`_)
 * if you use ePSFs: ``photutils`` (`Bradley et al. 2024 <https://zenodo.org/records/13989456>`_) and `Anderson & King (2000) <https://ui.adsabs.harvard.edu/abs/2000PASP..112.1360A/abstract>`_/`Anderson (2016) <https://ui.adsabs.harvard.edu/abs/2016wfc..rept...12A/abstract>`_
 * if you use STDPSFs: ``photutils`` (`Bradley et al. 2024 <https://zenodo.org/records/13989456>`_) and `Anderson (2016) <https://ui.adsabs.harvard.edu/abs/2016wfc..rept...12A/abstract>`_ + the relevant instrument paper -- MIRI (`Libralato et al. 2024 <https://ui.adsabs.harvard.edu/abs/2024PASP..136c4502L/abstract>`_); NIRISS (`Libralato et al. 2023 <https://ui.adsabs.harvard.edu/abs/2023ApJ...950..101L/abstract>`_) 
