+++ docs/psf.rst
@@ -5,7 +5,7 @@ Top-level Functions
 
 ``spike.psf`` is the primarily user-facing portion of the code. In most cases, you will only interact with ``spike.psf.hst``, ``spike.psf.jwst``, or ``spike.psf.roman``.
 
-The `quickstart guide <https://spike-psf.readthedocs.io/en/latest/quickstart.html>`_ offers a sense of the required inputs. See below for details of individual functions.
+The `quickstart guide <https://spike-psf.readthedocs.io/en/latest/quickstart.html>`_ and examples (for `ACS <https://spike-psf.readthedocs.io/en/latest/acs_example.html>`_, `WFC3 <https://spike-psf.readthedocs.io/en/latest/wfc3_example.html>`_, and `NIRCam <https://spike-psf.readthedocs.io/en/latest/nircam_example.html>`_) offer a sense of the required inputs. See below for details of individual functions.
 
 .. automodule:: spike.psf
 	:imported-members:
