+++ .github/ISSUE_TEMPLATE/bug-unexpected-behavior-report.md
@@ -1,0 +1,24 @@
+---
+name: Bug/unexpected behavior report
+about: Please report any bugs or unexpected behavior here.
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Description of the error or behavior:**
+Please include a screenshot or code block of any outputs.
+
+**Description of what you ran:**
+Please include a code block, showing the syntax you used to run the code, including any variable definitions. It is also helpful to know what data were used -- if you can link to MAST or similar (or offer instrument details and observation IDs), all the better. (If the data are not public, or you have concerns about being specific, please share as much as you can and/or share data that duplicates your concern.)
+
+**Description of your system/environment:**
+Please include a brief note, including what version of `python` and `spike` you are using, along with information about your OS. (Attaching a requirements.txt file here will be very useful. To do this, run `pip freeze > requirements.txt` from your command line in your working environment.)
+
+**Any other pertinent information:**
+
+(Please check all that apply.) **Before opening this issue, I have:**
+- [ ] checked the [documentation](https://spike-psf.readthedocs.io)
+- [ ] checked the [FAQs](https://github.com/avapolzin/spike/blob/master/FAQ.md)
+- [ ] checked [closed issues](https://github.com/avapolzin/spike/issues?q=is%3Aissue%20state%3Aclosed)
