### 2025-05-14 #8200 @editorialbot
[REVIEW]: spike: A tool to drizzle HST, JWST, and Roman point spread functions for improved analyses
**Submitting author:** @avapolzin (<a href="http://orcid.org/0000-0002-5283-933X">Ava Polzin</a>)
**Repository:** https://github.com/avapolzin/spike
**Branch with paper.md** (empty if default branch):
**Version:** 1.1.0
