### 2025-05-14 2afa1c9
fix for error raised in #6
