### 2025-05-21 #8200 @EdwardBerman
I think it's also important for Figures 1 and 2 in the paper to include the color bar to make the dynamic range (and likely log scaling) more apparent. I'm wondering about the PSFex model in Figure 2, I've found in one of my [papers](https://arxiv.org/abs/2401.11625v3) PSFex can produce pretty good models for JWST data. The config used to generate the model can likely be improved
