### 2025-05-23 9acb972
add test tolerance as in #7, address some comments raised in openjournals/joss-reviews #8200
