+++ paper/paper.md
@@ -19,7 +19,7 @@ bibliography: paper.bib
 
 # Summary
 
-Point spread functions (PSFs) describe the distribution of light for a pure point source in an astronomical image due to the instrument optics. For deconvolution, as for point source photometry and for source removal, it is key to have an accurate PSF for a particular image. Space-based telescopes can then pose a challenge as their PSFs are informed by their complex construction, and the myriad of pointings and rotations used to capture deep images. These telescopes also capture the highest resolution images of astronomical sources, resolving stars around even relatively distant galaxies. Proper co-addition of PSFs at a specific source position for space-based imaging is then both critical and challenging. This code, `spike`, generates model PSFs and runs them through the same processing pipeline used to derive deep, co-added images, providing correctly co-added and resampled PSFs for images from the _Hubble Space Telescope_, the _James Webb Space Telescope_, and the Nancy Grace Roman Space Telescope. 
+Instrumental point spread functions (PSFs) describe the distribution of light for a pure point source in an astronomical image due to the instrument optics. For deconvolution, as for point source photometry and for source removal, it is key to have an accurate PSF for a particular image. Space-based telescopes can then pose a challenge as their PSFs are informed by their complex construction, and the myriad of pointings and rotations used to capture deep images. These telescopes also capture the highest resolution images of astronomical sources, resolving stars around even relatively distant galaxies. Proper co-addition of PSFs at a specific source position for space-based imaging is then both critical and challenging. This code, `spike`, generates model PSFs and runs them through the same processing pipeline used to derive deep, co-added images, providing correctly co-added and resampled PSFs for images from the _Hubble Space Telescope_, the _James Webb Space Telescope_, and the Nancy Grace Roman Space Telescope. 
 
 
 # Statement of Need
