### 2025-05-23 71ebeb4
add test tolerance as in #7, address some comments raised in openjournals/joss-reviews #8200
