### 2025-05-23 dd1a9ac
add test tolerance as in #7, address some comments raised in openjournals/joss-reviews #8200
