+++ tests/test_outputs.py
@@ -71,72 +71,116 @@ import albumpl.cmap
 from astropy.io import fits
 
 
-fig, ax = plt.subplots(1, 5, figsize = (20, 4), gridspec_kw = {'wspace':0.1})
+def tick_formatter(y, pos):
+	if y <= 0:
+		if y == 0:
+			return 0
+		if y < 0:
+			label = str(np.log10(abs(y))).split('.')
+			exp = float(label[0])
+			if exp < 1:
+				pref = 10**float('-0.'+label[1])
+			if exp >= 1:
+				pref = 10**float('0.'+label[1])
+			if pref >= 10:
+				pref /= 10
+				exp += 1
+			if pref < 1:
+				exp -= 1
+			return r'$-%d \times 10^{%d}$'%(pref, exp)
+	else:
+		label = str(np.log10(abs(y))).split('.')
+		exp = float(label[0])
+		if exp < 1:
+			pref = 10**float('-0.'+label[1])
+		if exp >= 1:
+			pref = 10**float('0.'+label[1])
+		if pref >= 10:
+			pref /= 10
+			exp += 1
+		if pref < 1:
+			pref *= 10
+			exp -= 1
+		return r'$%i \times 10^{%i}$'%(pref, exp)
+
+
+fig, ax = plt.subplots(1, 5, figsize = (20, 5), gridspec_kw = {'wspace':0.1})
 
 tt = fits.open('psfs_tiny/150d08m15.267s+2d09m52.304s_F475W_psf_drz.fits')
 ttim = tt[1].data[325:465, 940:1080]
-ax[0].imshow(ttim, vmin = np.percentile(ttim, 20), vmax = np.percentile(ttim, 97), origin = 'lower', cmap = 'LondonCalling_r')
+cbar0 = ax[0].imshow(ttim, vmin = np.percentile(ttim, 20), vmax = np.percentile(ttim, 97), origin = 'lower', cmap = 'LondonCalling_r')
+plt.colorbar(cbar0, ax = ax[0], location = 'top', orientation = 'horizontal', pad = 0.05, format = tick_formatter)
 ax[0].text(10, 125, s = 'TinyTim', fontsize = 15, fontfamily = 'monospace', color = '#343530')
 ax[0].axis('off')
 
 ttg = fits.open('psfs_tinygillis/150d08m15.267s+2d09m52.304s_F475W_psf_drz.fits')
 ttgim = ttg[1].data[325:465, 940:1080]
-ax[1].imshow(ttgim, vmin = np.percentile(ttgim, 20), vmax = np.percentile(ttgim, 97), origin = 'lower', cmap = 'LondonCalling_r')
+cbar1 = ax[1].imshow(ttgim, vmin = np.percentile(ttgim, 20), vmax = np.percentile(ttgim, 97), origin = 'lower', cmap = 'LondonCalling_r')
+plt.colorbar(cbar1, ax = ax[1], location = 'top', orientation = 'horizontal', pad = 0.05, format = tick_formatter)
 ax[1].text(10, 125, s = 'TinyTim (Gillis+20)', fontsize = 15, fontfamily = 'monospace', color = '#343530')
 ax[1].axis('off')
 
 std = fits.open('psfs_stdpsf/150d08m15.267s+2d09m52.304s_F475W_psf_drz.fits')
 stdim = std[1].data[325:465, 940:1080]
 # ax[2].imshow(stdim, vmin = np.percentile(stdim[stdim != 0], 5), vmax = np.percentile(stdim[stdim != 0], 60), origin = 'lower', cmap = 'LondonCalling_r')
[…]
