### 2025-05-23 #8200 @avapolzin
Ah, realizing I didn't note it above, but I also added a direct link to the RTD to the repo's about section. The docs are also linked from the README and the setup.py metadata, so hopefully ample opportunity for people to know where to go. :)
