### 2025-05-23 f5ae6cc
change ADS abbreviations for full journal names
