### 2025-05-25 98aea94
docs/paper typo fixes
