### 2025-05-26 #8200 @EdwardBerman
Hi @avapolzin , thank you for this response. This answers my questions and concerns for those points.

I am trying to reproduce your example notebooks, however, the link keeps timing out on me. Either the link is invalid or the site is temporarily down. Regardless, I think it would be better if the necessary fits files are a part of the repo (if not too big) or some other service for easier and more direct download of the desired files.

Once I am able to get through these notebooks, I can check off the last 3 items on my checklist
