### 2025-05-26 a25647f
add link to Box folders with example data
