### 2025-05-26 #8200 @avapolzin
Thank you, @EdwardBerman, MAST can definitely be temperamental. I've uploaded the example data to a [Box folder](https://uchicago.box.com/s/1a0ip1q4dsdoswpi86630m02275z95j1) and just pushed a commit that adds that link to the example notebooks and the example README. (Most of the files are ~100-200 MB, so unfortunately, the external download is more feasible than including the data in the repo directly.)
