### 2025-05-27 6737231
fix tiny typo in the docs
