+++ docs/install.rst
@@ -43,7 +43,7 @@ If you install all of the optional dependencies described above, your shell's st
 	# export CRDS_SERVER_URL="https://jwst-crds.stsci.edu"
 	# export CRDS_SERVER_URL="https://roman-crds.stsci.edu"
 
-Since both JWST and Roman CRDS servers may be used, these variables are defined directly within ``spike.psf.jwst`` and ``spike.psf.roman`` and do not need to be added to your startup file. Similarly, ``spike`` doe not require that the ``tiny1``, ``tiny2``, and ``tiny3`` aliases are set up, but most ``TinyTim`` users will want to add these to their startup file regardless.
+Since both JWST and Roman CRDS servers may be used, these variables are defined directly within ``spike.psf.jwst`` and ``spike.psf.roman`` and do not need to be added to your startup file. Similarly, ``spike`` does not require that the ``tiny1``, ``tiny2``, and ``tiny3`` aliases are set up, but most ``TinyTim`` users will want to add these to their startup file regardless.
 
 Additionally, ``spike`` is written to be backwards compatible with ``WebbPSF`` installations.
 
