### 2025-05-27 d0d1fa7
fix another small typo in docs
