### 2025-05-27 f6d6f9f
add example link to psfgen.rst
