+++ docs/psfgen.rst
@@ -5,6 +5,8 @@ PSF Generation
 
 Functions included here relate to the creation of PSFs, which will later be drizzled by the HST, JWST, or Roman pipeline functions.
 
+These functions can also be used on their own to create single frame PSFs in python using consistent syntax. See `this example <https://spike-psf.readthedocs.io/en/latest/psf_generation.html>`_ for more details.
+
 .. automodule:: spike.psfgen
 	:imported-members:
 	:members:
