### 2025-05-30 #8200 @tobias-liaudat
Hi @avapolzin

Thanks for this contribution. This work will definitely be useful for the community and is very welcome for publication.

I found a typo in the statement of need, “`the resultant co-added PSF is not output,`" → “… is not optimal” .

Concerning relevant literature, there is a review on PSF models that (I wrote and) could be good for introducing the PSFs [(paper)](https://dx.doi.org/10.1088/1361-6420/acb664).

The article is well written. The only comment I have is that it would be good to add more details on the co-added models of the pipelines handled, JWST and Roman.

The code documentation looks good.

Great job with the FAQs document :) After downloading all the indicated files and finding one of the problems I had in the FAQs (the “ValueError: WCSNAME 'TWEAK' …”), I was able to run the `tests/`  without errors.

Concerning the example notebooks and the annoying use of environment variables inside them, I would add a suggestion for end users that might be easier than the current one. The suggestion is to manually set the environment variable(s) before importing spike by using

```python
import os
os.environ['TINYTIM'] = "path_to_tinytim"
```

I did this to run them, which I found to be an easier solution. In any case, the solution you proposed is totally fine, so consider this a suggestion :) I was able to run all the example notebooks and reproduce the results without errors.
