### 2025-05-30 5fbd0cf
make changes raised in openjournals/joss-reviews#8200
