+++ paper/paper.bib
@@ -64,76 +64,6 @@ archivePrefix = {arXiv},
 }
 
 
-@ARTICLE{Rogowski.etal.2023,
-  author={Rogowski, Marcin and Aseeri, Samar and Keyes, David and Dalcin, Lisandro},
-  journal={IEEE Transactions on Parallel and Distributed Systems}, 
-  title={mpi4py.futures: MPI-Based Asynchronous Task Execution for Python}, 
-  year={2023},
-  volume={34},
-  number={2},
-  pages={611-622},
-  keywords={Python;Task analysis;Message systems;Libraries;Standards;Codes;Sockets;MPI;Python;parallelism;master-worker;parallel programming models;distributed computing;high performance computing;task execution;multiprocessing},
-  doi={10.1109/TPDS.2022.3225481}
-}
-
-@ARTICLE{Dalcin.Fang.2021,
-  author={Dalcin, Lisandro and Fang, Yao-Lung L.},
-  journal={Computing in Science & Engineering}, 
-  title={mpi4py: Status Update After 12 Years of Development}, 
-  year={2021},
-  volume={23},
-  number={4},
-  pages={47-54},
-  keywords={Python;Graphics processing units;High performance computing;Semantics;Programming;Task analysis},
-  doi={10.1109/MCSE.2021.3083216}
-}
-
-@article{Dalcin.etal.2011,
-title = {Parallel distributed computing using Python},
-journal = {Advances in Water Resources},
-volume = {34},
-number = {9},
-pages = {1124-1139},
-year = {2011},
-note = {New Computational Methods and Software Tools},
-issn = {0309-1708},
-doi = {https://doi.org/10.1016/j.advwatres.2011.04.013},
-url = {https://www.sciencedirect.com/science/article/pii/S0309170811000777},
-author = {Lisandro D. Dalcin and Rodrigo R. Paz and Pablo A. Kler and Alejandro Cosimo},
-keywords = {Python, MPI, PETSc},
-abstract = {This work presents two software components aimed to relieve the costs of accessing high-performance parallel computing resources within a Python programming environment: MPI for Python and PETSc for Python. MPI for Python is a general-purpose Python package that provides bindings for the Message Passing Interface (MPI) standard using any back-end MPI implementation. Its facilities allow parallel Python programs to easily exploit multiple processors using the message passing paradigm. PETSc for Python provides access to the Portable, Extensible Toolkit for Scientific Computation (PETSc) libraries. Its facilities allow sequential and parallel Python applications to exploit state of the art algorithms and data structures readily available in PETSc for the solution of large-scale problems in science and engineering. MPI for Python and PETSc for Python are fully integrated to PETSc-FEM, an MPI and PETSc based parallel, multiphysics, finite elements code developed at CIMEC laboratory. This software infrastructure supports research activities related to simulation of fluid flows with applications ranging from the design of microfluidic devices for biochemical analysis to modeling of large-scale stream/aquifer interactions.}
-}
-
-@article{Dalcin.etal.2008,
-title = {MPI for Python: Performance improvements and MPI-2 extensions},
-journal = {Journal of Parallel and Distributed Computing},
-volume = {68},
-number = {5},
-pages = {655-662},
-year = {2008},
-issn = {0743-7315},
-doi = {https://doi.org/10.1016/j.jpdc.2007.09.005},
[…]
