+++ docs/install.rst
@@ -101,4 +101,16 @@ If you are still using ``WebbPSF`` instead of ``STPSF``, your key/value pair wil
 
 In my experience, this works for notebooks instantiated from the command line or an IDE, but does not consistently work with, e.g., the JupyterLab application. See also `this discussion <https://stackoverflow.com/questions/37890898/how-to-set-env-variable-in-jupyter-notebook>`_ for other ways to set up environment variables with Jupyter kernels.
 
+One alternative that discussion raises, is to use ``os`` to set environment variables before importing ``spike`` as follows:
+
+.. code-block:: python
+
+	import os
+
+	os['TINYTIM'] = '/path/to/tinytim'
+	os['STPSF_PATH'] = 'path/to/stpsf'
+	os['CRDS_PATH'] = 'path/to/crds/cache'
+
+	import spike
+
 If, instead, you have installed ``spike`` to your root environment, you can simply use your standard python kernel.
