### 2025-05-30 37bf7ff
make changes raised in openjournals/joss-reviews#8200
