### 2025-05-30 #8200 @avapolzin
Thank you very much, @tobias-liaudat!

I have added the paper you suggested and specifically given instructions for manually setting the environment variables in both the example_notebooks README and in the RTD. I have also changed the language you pointed out -- I had intended to use the word "output" there, as the drizzled PSFs are not returned by the PSF photometry process, but I have clarified that the "resultant PSF is not available to users".

There were initially more details of the way that the `jwst` and `romancal` pipelines are incorporated here, but I removed that discussion to get closer to the word limit. As it is I believe the paper is at ~1200 words. I can certainly add that discussion back if it's preferred (probably mostly a question for @ivastar).
