### 2025-05-30 #8200 @EdwardBerman
Seconding what @tobias-liaudat , the environment variables were an obstacle for me getting through the notebooks. With this change it should be easier for me to get through the rest of the checklist 😄
