### 2025-05-30 avapolzin/spike#8 @marcio-melendez
Imrpovements to  jwpsf function in psfgen.py
A couple of comments in regards to the use of STPSF in the code, in partcular under the jwpsf function

1. There is a convenient function in STPSF that allows for the simulation to do the setup from an observation, setup_sim_to_match_file. In this way you don’t have to setup the instrument by hand or in a couple of steps as it is in the current implementation. In addition, the current use of stpsf doesn’t use the fact that the optical quality of the telescope changes over time, some of these changes can be more evident at shorter wavelengths. In other words, it seems that the STPSF simulation use the nominal in-flight phase map but it should be using the phase map closer to the actual observation. See [https://stpsf.readthedocs.io/en/latest/jwst_measured_opds.html](https://stpsf.readthedocs.io/en/latest/jwst_measured_opds.html)

I know that adding setup_sim_to_match_file may involve a bit more of coding, but I think it may be an important addition. Alternatively, the same way that you read all the instrument parameters you could also read the date of the observation and then just load the phase map for that day with the function load_wss_opd_by_date

2. Currently, the default value of the PSF sampling when calling STPSF is = 1. Although this is a possible value, I think in this case it is incorrect. To properly simulate a PSF you need to have the correct oversampled factor in order to avoid numerical artifacts, see:  https://stpsf.readthedocs.io/en/latest/sampling.html#per-instrument-sampling-requirements
In addition, note that ext = 3 from the STPSF simulation will always give you the detector sampled PSFs, independent of the oversampling. So, there is not need to set a “sample” value equal to 1 in order to get a detector sampled PSF or to resampled the PSF to detector pixel scale. Also, I would suggest to call the name of the extension explicitly, for example “DET_DIST“ Instead of 3, this helps a bit with adding some context in the code. As a reminder, here are the definitions of all the STPSF extensions:
[https://stpsf.readthedocs.io/en/latest/jwst_matching_psfs_to_data.html#A-reminder-about-output-data-products](https://stpsf.readthedocs.io/en/latest/jwst_matching_psfs_to_data.html#A-reminder-about-output-data-products)
