### 2025-05-30 avapolzin/spike#9 @obi-wan76
Notebook nircam_example.ipynb
I was able to run the entire notebook without issues. Here is a list of my comments and suggestions.

- Change for clarity.  “This notebook uses observations the following files:” to “This notebook uses observations from the following files:” . You could also add a note about the program that you are using for this notebook. For example, “For this example we are going to use observations obtained by NIRCam filter F115W  from “PANORAMIC -- A Pure Parallel Wide Area Legacy Imaging Survey at 1-5 Micron” PI: Dr. Christina Williams. There are 3 exposures in this visit.

- I think this notebook could benefit from some sections. It is hard to follow everything as a single document. Example of sections, 1) initial setup and visualization, 2) PSF generation methods (sub section STDPSF)

- I think it is fine to do an index array selection to show the PSFs but I would recommend the use of Cutout2D, https://docs.astropy.org/en/stable/api/astropy.nddata.Cutout2D.html You can center on a pixel or WCS coordinate, it can make comparisons easier.

- Replace all the WebbPSF to STPSF , e.g., “We'll do this using the default WebbPSF parameters below.” Do the change when appropriate, mostly in the text,  because I know there are some webbpsf usage intended to avoid confusion.
