### 2025-05-30 #8200 @obi-wan76
Hi @avapolzin,
I started my review with a couple of  new issues:

- [issue #8](https://github.com/avapolzin/spike/issues/8)
- [issue #9](https://github.com/avapolzin/spike/issues/9)

Overall, this is a valuable tool to have. Congratulation on all the effort that you have put on the development so far and for making an effort to share it with the community.

Some additional comments that  are not necessarily labeled as issues:

I copy/paste the coordinates for the object use in this notebook, nircam_example.ipynb, into the actual NIRCam image (e.g., 10:00:31.4320, +2:10:26.290) and there is no source in the image. In fact, the image doesn’t have any PSFs, so I don’t understand the goal of this exercise.
In other words, what’s the purpose of the actual images beside getting the instrumental setup? From the notebook it is unclear what’s the science goal that we are trying to accomplish, for example, to create a drizzle PSF that we can use to do PSF subtraction on the actual observations. Or how does this simulated PSFs compared to the actual Level 3 products of the pipeline when combing dithers or how does compare to the semi-empirical ePSF? I think it could be useful to have a science case in the notebook so people can have a better understanding on why (or if) they should be using this tool for their science.

I am happy to follow up any comments either here and/or on the issues.

Thanks!
