### 2025-06-03 avapolzin/spike#8 @avapolzin
Thank you, @obi-wan76 -- I've switched to using `stpsf.setup_sim_to_match_file`, changed 3 -> 'DET_DIST' in the STPSF HDU indexing, and adopted a default PSF sampling value of 4. To make that worthwhile for users, I've also added an argument that allows for the full STPSF output to be saved from `spike.psfgen`.
