### 2025-06-03 d41f95d
oversample default int
