### 2025-06-03 79af340
reorder spike.jwpsf arguments
