### 2025-06-03 76834df
update to tools.rewrite_fits to handle PSFs near the edges of frames
