### 2025-06-03 db201f4
update tools.cutout to handle PSFS near the edges of frames
