### 2025-06-04 dacc8cd
make change to notebooks based on #9
