+++ spike/psf/psf.py
@@ -20,7 +20,7 @@ warnings.formatwarning = warning_on_one_line
 
 
 def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermethod = None, 
-		savedir = 'psfs', drizzleimgs = False, pretweaked = False,
+		savedir = 'psfs', drizzleimgs = False, objonly = True, pretweaked = False,
 		keeporig = True, plot = False, verbose = False, parallel = False, out = 'fits', 
 		tweakparams = {'threshold':6.0, 
 					   'searchrad':3.0, 
@@ -69,6 +69,7 @@ def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermetho
 				headers are from the original images (see spike.tools.rewrite_fits, which can be used to this end).
 		savedir (str): Where the PSF models and drizzled PSF will be saved. Defaults to 'psfs'.
 		drizzleimgs (bool): If True, will drizzle the input images at the same time as creating a drizzled psf.
+		objonly (bool): If True, only drizzles input images that cover the selected obj.
 		pretweaked (bool): If True, skips TweakReg steps to include fine WCS corrections.
 		keeporig (bool): If True (and pretweaked = False), create copy of img_dir before TweakReg.
 		plot (bool): If True, saves .pngs of the model PSFs.
@@ -167,9 +168,11 @@ def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermetho
 			tweakreg.TweakReg(filelist[fk], **tweakparams)
 
 	drizzlelist = {} #write file prefixes to drizzle per object per filter
+	imglist = {} #images to drizzle per object per filter (used if objonly = True)
 	if genpsf: #generate model PSFs for each image + object
 		if type(obj) == str: #check number of objects
 			drizzlelist[obj] = {}
+			imglist[obj] = {}
 			skycoords = tools.objloc(obj)
 			for i in imgs:
 				pos = tools.checkpixloc(skycoords, i, inst, camera)
@@ -180,11 +183,14 @@ def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermetho
 				if skycoords.dec.deg < 0:
 					coordstring += str(skycoords.dec)
 
-				modname = i.replace('.fits', '_'+coordstring+'_%s'%pos[3]+'_topsf.fits')
+				modname = i.replace('%s.fits'%img_type, coordstring+'_%s'%pos[3]+'_topsf_%s.fits'%img_type)
 				if np.isfinite(pos[0]): #confirm object falls onto image
 					if pos[3] not in drizzlelist[obj].keys():
 						drizzlelist[obj][pos[3]] = []
+						imglist[obj][pos[3]] = []
 					drizzlelist[obj][pos[3]].append(modname)
+					imglist[obj][pos[3]].append(i)
+
 
 					psffunc(skycoords, i, imcam, pos, plot, verbose, **kwargs)
 
@@ -193,6 +199,7 @@ def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermetho
 
 			for o in obj:
 				drizzlelist[o] = {}
+				imglist[o] = {}
 				skycoords.append(tools.objloc(o))
 			
 			for i in imgs:
@@ -211,11 +218,13 @@ def hst(img_dir, obj, img_type, inst, camera = None, method='TinyTim', usermetho
 						if coord.dec.deg < 0:
 							coordstring += str(coord.dec)
 
-						modname = i.replace('.fits', '_'+coordstring+'_%s'%pos[3]+'_topsf.fits')
[…]
