### 2025-06-12 #8200 @obi-wan76
@avapolzin the changes looks good to me, thanks for addressing my comments. One final comments, in the paper, can you change the WebbPSF label in first panel of the second figure to  STPSF. Thanks!
I am done with my review.

@ivastar
