### 2025-06-12 #8200 @avapolzin
Since I'm looking at this thread -- @EdwardBerman, just want to confirm that the data download from Box worked for you. (Sounds like it worked for others.)
