### 2025-06-12 #8200 @obi-wan76
> Thank you, [@obi-wan76](https://github.com/obi-wan76)! I've changed the label in the figure to "STPSF (WebbPSF)" just to remain consistent with other mentions in the body of the paper. If you prefer I simply use "STPSF", I can make that further update.

For the label, I think just STPSF is better. Thanks!
