### 2025-06-12 498cc45
fix typos in example notebook markdown
