### 2025-06-12 #8200 @ivastar
@avapolzin there seems to be an issue with the bibtex. I think the problem is with the percent signs in front some of the entries, maybe delete these?

>         %   eid = {arXiv:2209.09253},
>         % pages = {arXiv:2209.09253},

> Checking the BibTeX entries failed with the following error:
>
> ```
> Failed to parse BibTeX on value "eid" (NAME) [#<BibTeX::Bibliography data=[1]>, "@", #<BibTeX::Entry >, {:author=>["{Mandelbaum}, Rachel and {Jarvis}, Mike and {Lupton}, Robert H. and {Bosch}, James and {Kannawadi}, Arun and {Murphy}, Michael D. and {Zhang}, Tianqing and {the LSST Dark Energy Science Collaboration}"], :title=>["{PSFs of coadded images}"], :journal=>["Open Journal of Astrophysics"], :keywords=>["Astrophysics - Instrumentation and Methods for Astrophysics"], :year=>"2023", :month=>[:feb], :volume=>["6"]}, ",", "%"]
> ```
