+++ paper/paper.bib
@@ -37,8 +37,8 @@
          year = 2023,
         month = feb,
        volume = {6},
-        %   eid = {arXiv:2209.09253},
-        % pages = {arXiv:2209.09253},
+           eid = {arXiv:2209.09253},
+         pages = {arXiv:2209.09253},
           doi = {10.21105/astro.2209.09253},
 archivePrefix = {arXiv},
        eprint = {2209.09253},
