+++ paper/paper.bib
@@ -482,9 +482,7 @@ year = {2024}
   journal = {GitHub repository},
   url = {https://github.com/spacetelescope/romancal}
 }
-%authors for these two taken from pyproject.toml
-%dates are from most recent update? could also do copyright dates
-%which would be 2010 and 2020 respectively
+
 @misc{stcal,
   author = {STScI},
   title = {STCAL: STScI Calibration algorithms and tools.},
