### 2025-06-12 db06ece
update to 1.1.0 - version with all changes from JOSS review
