### 2025-06-12 0288462
fix list syntax for rtd
