+++ docs/changelog.rst
@@ -7,68 +7,80 @@ See below for a record of salient changes to the code itself. Updates to e.g. do
 
 
 **v1.1.0 (June 12, 2025)**
+
 Version that incorporates all changes from JOSS review.
-- Move file suffixes in _topsf files to generate drc/drz files based on inputs
-- Update naming scheme in `spike.tools.rewrite_fits`
-- Bug fix in renaming user-generated PSF models
+
+* Move file suffixes in _topsf files to generate drc/drz files based on inputs
+* Update naming scheme in `spike.tools.rewrite_fits`
+* Bug fix in renaming user-generated PSF models
 
 
 **v1.0.9 (June 2, 2025)**
-- Add tolerance for sub-pixel coordinate differences in tests
-- Update default sampling for `STPSF`/`WebbPSF`
-- Adopt auto setup of model from img
-- Use explicit indexing of HDU in `spike.psfgen.jwpsf`
-- Fix "savefull" typo in `spike.psfgen.jwpsf`
-- Add support for PSFs near frame edges to `spike.tools.rewrite_fits` and `spike.tools.cutout`
+
+* Add tolerance for sub-pixel coordinate differences in tests
+* Update default sampling for `STPSF`/`WebbPSF`
+* Adopt auto setup of model from img
+* Use explicit indexing of HDU in `spike.psfgen.jwpsf`
+* Fix "savefull" typo in `spike.psfgen.jwpsf`
+* Add support for PSFs near frame edges to `spike.tools.rewrite_fits` and `spike.tools.cutout`
 
 
 **v1.0.8 (May 14, 2025)**
-- Fix path to "removedir" when "finalonly = True"
+
+* Fix path to "removedir" when "finalonly = True"
 
 
 **v1.0.7 (May 12, 2025)**
-- Update output filename post-drizzling in `spike.psf`
-- Add option to only retain drizzled products ("finalonly = True")
-- Add "maskval" keyword argument as user option for `spike.psfgen.effpsf`
-- Adopt nanmedian instead of median for masked data in same module
-- Change STDPSF link format for WFC3/UVIS
-- Add support for floats in extension versions in `spike.tools.rewrite_fits`
-- Switch to `STPSF` as default from `WebbPSF`; retain support for `WebbPSF`
-- Fix bug in `astropy.wcs.utils.pixel_to_skycoords` call for some cutouts
+
+* Update output filename post-drizzling in `spike.psf`
+* Add option to only retain drizzled products ("finalonly = True")
+* Add "maskval" keyword argument as user option for `spike.psfgen.effpsf`
+* Adopt nanmedian instead of median for masked data in same module
+* Change STDPSF link format for WFC3/UVIS
+* Add support for floats in extension versions in `spike.tools.rewrite_fits`
+* Switch to `STPSF` as default from `WebbPSF`; retain support for `WebbPSF`
+* Fix bug in `astropy.wcs.utils.pixel_to_skycoords` call for some cutouts
 
 
 **v1.0.6 (May 8, 2025)**
-- Add example notebooks
-- Force format input path
-- Update `spike.psf` to use same "modname" everywhere
-- Fix unnecessary "+" in filenames for objects with negative declination
-- Re-index loops across multiple files
+
+* Add example notebooks
+* Force format input path
+* Update `spike.psf` to use same "modname" everywhere
+* Fix unnecessary "+" in filenames for objects with negative declination
+* Re-index loops across multiple files
 
 
 **v1.0.5 (April 2, 2025)**
-- Add explicit extension number in `spike.tools.rewrite_fits`
-- Fix bug that resulted in declination being duplicated in output filenames
+
+* Add explicit extension number in `spike.tools.rewrite_fits`
[…]
