### 2025-06-12 6d80b54
fix Liaudat+2023 reference
