### 2025-06-12 200d4aa
arxiv -> fspas version
