+++ paper/paper.bib
@@ -783,20 +783,21 @@ version = {2.0.0}
 }
 
 @ARTICLE{Liaudat.etal.2023,
-       author = {{Liaudat}, Tob{\'\i}as I. and {Starck}, Jean-Luc and {Kilbinger}, Martin},
-        title = "{Point spread function modelling for astronomical telescopes: a review focused on weak gravitational lensing studies}",
-      journal = {Frontiers in Astronomy and Space Sciences},
-     keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Computer Science - Computer Vision and Pattern Recognition},
-         year = 2023,
-        month = oct,
-       volume = {10},
-          eid = {1158213},
-        pages = {1158213},
-          doi = {10.3389/fspas.2023.1158213},
-archivePrefix = {arXiv},
-       eprint = {2306.07996},
- primaryClass = {astro-ph.IM},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023FrASS..1058213L},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
+  
+AUTHOR={Liaudat, Tobías I.  and Starck, Jean-Luc  and Kilbinger, Martin },
+         
+TITLE={Point spread function modelling for astronomical telescopes: a review focused on weak gravitational lensing studies},
+        
+JOURNAL={Frontiers in Astronomy and Space Sciences},
+        
+VOLUME={Volume 10 - 2023},
 
+YEAR={2023},
+
+URL={https://www.frontiersin.org/journals/astronomy-and-space-sciences/articles/10.3389/fspas.2023.1158213},
+
+DOI={10.3389/fspas.2023.1158213},
+
+ISSN={2296-987X},
+
+ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of paramount importance in astronomical observations, as it allows for the correction of distortions and blurring caused by the telescope and atmosphere. PSF modelling is crucial for accurately measuring celestial objects’ properties. The last decades have brought us a steady increase in the power and complexity of astronomical telescopes and instruments. Upcoming galaxy surveys like <italic>Euclid</italic> and Legacy Survey of Space and Time (LSST) will observe an unprecedented amount and quality of data. Modelling the PSF for these new facilities and surveys requires novel modelling techniques that can cope with the ever-tightening error requirements. The purpose of this review is threefold. Firstly, we introduce the optical background required for a more physically motivated PSF modelling and propose an observational model that can be reused for future developments. Secondly, we provide an overview of the different physical contributors of the PSF, which includes the optic- and detector-level contributors and atmosphere. We expect that the overview will help better understand the modelled effects. Thirdly, we discuss the different methods for PSF modelling from the parametric and non-parametric families for ground- and space-based telescopes, with their advantages and limitations. Validation methods for PSF models are then addressed, with several metrics related to weak-lensing studies discussed in detail. Finally, we explore current challenges and future directions in PSF modelling for astronomical telescopes.</p>}}
