### 2025-06-12 bd6f23d
update changelog language
