### 2025-06-12 c076da7
add clarifying language to changelog
