### 2025-06-27 269a627
update WFC3 STDPSF per #11
