### 2025-06-27 1024bdb
update changelog to reflect WFC3 STDPSF update
