### 2025-07-02 #8200 @ivastar
Hi Ava,

Finally got around the reading the text and checking the references. Here are some comments.

References:
- Remove the PSFEx citation of ASCL - the correct reference is the Bertin 2011 paper.
- Remove the reference to the Gillis BitBucket repo - the correct reference is the paper and the repo is linked from that.
- The correct reference for the Drizzlepac handbook is: Hoffmann, S. L., Mack, J., et al., 2021, “The DrizzlePac Handbook”, Version 2.0, (Baltimore:STScI).
- There are two Perrin et al. citations that don't have dates, they appear as "n.d.-a" and "n.d.-b". I guess that's "no date" but it looks really wonky. Maybe add the years of the releases of the versions you specifically cite?

I opened issues on several (9?) of the other repos that don't have citation recommendations. This is out of the scope here.

Title: Please capitalize

Line 6: Instrumental point spread functions (PSFs) describe --> maybe do this in singular: "An instrumental point spread function (PSF) describes"
Line 7: instrument optics - > the optics of the instrument
Line 7-8: maybe reverting the second sentence would make it easier to read: An accurate PSF is key for deconvolution, point source photometry and source removal.
Line 9: as --> because; informed --> influenced
Line 12: stars around --> stars in
Line 13: This code --> The library described in this work
Line 19: Maybe plural would be better for this sentence? "The PSFs ... are" because it is not a single PSF (also a bunch of the comments in this paragraph are just switching to plural)
Line 19: "of generic interest to both " --> "of interest to users of both"
Line 20-21:  Start of a new sentence: PSFs can be impacted by **different** .... and **may impact** the analysis **results based on the data**
Line 22: PSF-PSFs
Line 23: image --> images; a PSF --> the PSFs
Line 27: This drizzled --> "However, this drizzled"
Line 28: for the source position --> for the positions of sources
Line 31: Instead --> alternatively; taken --> created
Line 34: remove "real"
Line 39: I think `grizli` should be lower case
Line 42: I would remove "onerous", maybe challenging or difficult; remove the quotations on realistic
Line 47: I would remove the footnote to `spike`. The final paper will contain a link to the repository.
Line 50, 54: replace .fits with capital FITS
Line 57: maybe consolidate the `stpsf` and `WebbPSF` citations all together so the text it not broken up?
Line 58: effective PSFs --> "ePSF" since this is how you refer to it in the figure captions
Line 73-74: the working directory need not be local --> this seems best included in the documentation
Line 76: `stpsf` now supports Roman observations (I think), will there be an update?

Figure 1 & 2: Why are the stretches different for the different panels? Is it possible to make them all the same?
