### 2025-07-02 0f52eb0
update paper.md and paper.bib based on openjournals/joss-reviews#8200
