+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '`spike`: A tool to drizzle _HST_, _JWST_, and Roman PSFs for improved analyses'
+title: '`spike`: A Tool to Drizzle _HST_, _JWST_, and Roman PSFs for Improved Analyses'
 tags:
   - Python
   - astronomy
@@ -13,33 +13,32 @@ authors:
 affiliations:
  - name: Department of Astronomy and Astrophysics, The University of Chicago, USA
    index: 1
-date: 5 December 2024
+date: 2 July 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-Instrumental point spread functions (PSFs) describe the distribution of light for a pure point source in an astronomical image due to the instrument optics. For deconvolution, as for point source photometry and for source removal, it is key to have an accurate PSF for a particular image. Space-based telescopes can then pose a challenge as their PSFs are informed by their complex construction, and the myriad of pointings and rotations used to capture deep images. These telescopes also capture the highest resolution images of astronomical sources, resolving stars around even relatively distant galaxies. Proper co-addition of PSFs at a specific source position for space-based imaging is then both critical and challenging. This code, `spike`, generates model PSFs and runs them through the same processing pipeline used to derive deep, co-added images, providing correctly co-added and resampled PSFs for images from the _Hubble Space Telescope_, the _James Webb Space Telescope_, and the Nancy Grace Roman Space Telescope. 
+An instrumental point spread function (PSF) describes the distribution of light for a pure point source in an astronomical image due to the optics of the instrument. An accurate PSF is key for deconvolution, point source photometry and source removal. Space-based telescopes can then pose a challenge because their PSFs are influenced by their complex construction, and the myriad of pointings and rotations used to capture deep images. These telescopes also capture the highest resolution images of astronomical sources, resolving stars in even relatively distant galaxies. Proper co-addition of PSFs at a specific source position for space-based imaging is then both critical and challenging. The library described in this work, `spike`, generates model PSFs and runs them through the same processing pipeline used to derive deep, co-added images, providing correctly co-added and resampled PSFs for images from the _Hubble Space Telescope_, the _James Webb Space Telescope_, and the Nancy Grace Roman Space Telescope. 
 
 
 # Statement of Need
 
[…]
