+++ paper/paper.bib
@@ -101,17 +101,10 @@ author = {Perrin, Marshall and Melendez, Marcio and Osborne, Shannon and Geda, R
 license = {BSD-3-Clause},
 title = {{WebbPSF}},
 url = {https://github.com/spacetelescope/webbpsf},
-version = {1.4.0}
+version = {1.4.0},
+year = 2024
 }
 
-@misc{Gillis.2019,
-  author = {Bryan Gillis},
-  title = {TinyTim_PSFs: Public repository for code to assist in generating PSFs with TinyTim.},
-  year = {2019},
-  publisher = {BitBucket},
-  journal = {BitBucket repository},
-  url = {https://bitbucket.org/brgillis/tinytim_psfs/src/master/}
-}
 
 @ARTICLE{Gillis.etal.2020,
        author = {{Gillis}, Bryan R. and {Schrabback}, Tim and {Marggraf}, Ole and {Mandelbaum}, Rachel and {Massey}, Richard and {Rhodes}, Jason and {Taylor}, Andy},
@@ -310,18 +303,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-
-
-@software{Bertin.2013,
-       author = {{Bertin}, Emmanuel},
-        title = "{PSFEx: Point Spread Function Extractor}",
- howpublished = {Astrophysics Source Code Library, record ascl:1301.001},
-         year = 2013,
-        month = jan,
-          eid = {ascl:1301.001},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ascl.soft01001B},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
 
 
 @ARTICLE{Troxel.etal.2023,
@@ -779,7 +760,8 @@ author = {Perrin, Marshall and Melendez, Marcio and Osborne, Shannon and Geda, R
 license = {BSD-3-Clause},
 title = {{STPSF}},
 url = {https://github.com/spacetelescope/stpsf},
-version = {2.0.0}
+version = {2.0.0},
+year = 2025
 }
 
 @ARTICLE{Liaudat.etal.2023,
