### 2025-07-02 #8200 @avapolzin
Thank you, @ivastar! I will follow up with a Zenodo DOI momentarily. I addressed all your comments above with the exception of using the same stretch for the output PSF model images in Figures 1 and 2, as the individual stretches are chosen based on the models and how they handle the background, the wings of the PSF, faint features, etc. I can also clarify this in the text if it's preferred.

There were two other points I wasn't able to meaningfully address as they didn't seem to require substantial changes on my end:
1.  I believe I'm already using the `Drizzlepac` reference that you suggest, as I grabbed the BibTeX from their [documentation](https://drizzlepac.readthedocs.io/en/latest/getting_started/citing_drizzlepac.html) and there is no other Hoffmann et al. (2021) entry in ADS. Happy to make further changes if I should still be amending that reference in some way, though.
2. On your last comment about `stpsf` already handling Roman observations, I've amended the paper text to clarify that I will update the *reduction* pipeline modules (for tweaking and resampling) to reflect the most updated software in use once Roman launches. I suppose it's an implicit assumption in this that `romancal` will change with on-sky data, so I can also clarify this further in the text if the changes to the paper to now are insufficient.
