### 2025-07-02 592e653
add Zenodo link to README
