### 2025-07-02 #8200 @avapolzin
Hi, @ivastar -- here is the link to the release [on GitHub](https://github.com/avapolzin/spike/releases/tag/v1.1.0) and [on Zenodo](https://zenodo.org/records/15791925).
