### 2025-07-02 6412fbb
setup.py consistent with v1.1.0
