### 2025-07-02 be001b0
address stretch in Figure 1 caption
