+++ paper/paper.md
@@ -34,7 +34,7 @@ For analyses to be consistent, mock PSFs may be used, but the proper treatment o
 
 Given a directory containing reduced and calibrated, but not yet co-added, FITS files from _HST_, _JWST_, or Roman and the coordinates of an object of interest, detector/chip-specifc model PSFs are directly "drizzled" by `spike`. A model PSF is generated for each unique input image and coordinate combination. Instrument information, including camera, filter, and, if necessary, chip, is automatically read from the header of each FITS file.
 
-![Comparison of drizzled PSFs generated for _HST_ images using the default parameters for different methods included in `spike`. All panels use the same ACS/WFC imaging of the COSMOS field in F475W. Note that the ePSF panel (second from right) shows some artifacts; the robustness of the effective PSF method is heavily dependent on the number of stars in the chosen field and may be changed by altering the star detection threshold. \label{fig:hstcomparison}](spike_psf_hstcompare.png)
+![Comparison of drizzled PSFs generated for _HST_ images using the default parameters for different methods included in `spike`. All panels use the same ACS/WFC imaging of the COSMOS field in F475W. The stretch of each model's colormap is chosen based on their individual hnadling of the background, wings, and faint features of the PSF. Note that the ePSF panel (second from right) shows some artifacts; the robustness of the effective PSF method is heavily dependent on the number of stars in the chosen field and may be changed by altering the star detection threshold. \label{fig:hstcomparison}](spike_psf_hstcompare.png)
 
 `spike` works with both model and empirical PSFs. The built-in PSF generation options are `TinyTim` [@Krist.etal.2011] and the @Gillis.etal.2020 modification, `STPSF` -- formerly `WebbPSF` [@Perrin_STPSF; @Perrin.etal.2012; @Perrin.etal.2014; @Perrin_WebbPSF], `photutils` ePSFs [@Anderson.King.2000; @Anderson.2016; @Bradley.etal], `PSFEx` [@Bertin.2011], and Space Telescope Science Institute's library of empirical STDPSFs [^2] [@Anderson.2016; @Libralato.etal.2023; @Libralato.etal.2024], all of which are included for having the ability to generate a model PSF for an arbitrary detector location. \autoref{fig:hstcomparison} compares output drizzled PSFs from the different methods for _HST_ and \autoref{fig:jwstcomparison} compares output drizzled PSFs for _JWST_. 
 
