### 2025-07-02 #8200 @avapolzin
Thanks, @ivastar, I've added mention of the different colormap stretch to the figure caption. On the point about `stpsf` -- yes, all of the infrastructure exists to use `spike` with Roman PSFs via `spike.psf.roman` using `stpsf` and the modified `romancal` pipeline in `spike.roman`. This was initially made explicit in the documentation, but I removed the mention to get closer to the word limit, as what exists is preliminary and will be updated to reflect the pipeline and its handling of on-sky data once Roman launces.
