### 2025-07-03 42965cb
update reference to romancal in paper
