### 2025-07-09 #8200 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
