+++ paper/paper.bib
@@ -548,6 +548,21 @@ year = {2024}
   url = {https://github.com/spacetelescope/stpipe}
 }
 
+@ARTICLE{Greenfield.etal.2015,
+       author = {{Greenfield}, P. and {Droettboom}, M. and {Bray}, E.},
+        title = "{ASDF: A new data format for astronomy}",
+      journal = {Astronomy and Computing},
+     keywords = {FITS, File formats, Standards, World coordinate system},
+         year = 2015,
+        month = sep,
+       volume = {12},
+        pages = {240-251},
+          doi = {10.1016/j.ascom.2015.06.004},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2015A&C....12..240G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
 @software{graham.etal,
   author       = {Brett Graham and
                   Dan D'Avella and
@@ -616,14 +631,20 @@ year = {2024}
   url = {https://github.com/yaml/pyyaml}
 }
 
-@misc{crds,
-  author = {{STScI CRDS s/w developers}},
-  title = {{crds}: Calibration Reference Data System, {HST/JWST/Roman} reference file management},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/crds}
+@ARTICLE{Greenfield.Miller.2016,
+       author = {{Greenfield}, P. and {Miller}, T.},
+        title = "{The Calibration Reference Data System}",
+      journal = {Astronomy and Computing},
+     keywords = {Calibration, Pipelines},
+         year = 2016,
+        month = jul,
+       volume = {16},
+        pages = {41-53},
+          doi = {10.1016/j.ascom.2016.04.001},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2016A&C....16...41G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
 
 @software{jsonschema,
   author       = {Julian Berman and
@@ -782,3 +803,18 @@ DOI={10.3389/fspas.2023.1158213},
 ISSN={2296-987X},
 
 ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of paramount importance in astronomical observations, as it allows for the correction of distortions and blurring caused by the telescope and atmosphere. PSF modelling is crucial for accurately measuring celestial objects’ properties. The last decades have brought us a steady increase in the power and complexity of astronomical telescopes and instruments. Upcoming galaxy surveys like <italic>Euclid</italic> and Legacy Survey of Space and Time (LSST) will observe an unprecedented amount and quality of data. Modelling the PSF for these new facilities and surveys requires novel modelling techniques that can cope with the ever-tightening error requirements. The purpose of this review is threefold. Firstly, we introduce the optical background required for a more physically motivated PSF modelling and propose an observational model that can be reused for future developments. Secondly, we provide an overview of the different physical contributors of the PSF, which includes the optic- and detector-level contributors and atmosphere. We expect that the overview will help better understand the modelled effects. Thirdly, we discuss the different methods for PSF modelling from the parametric and non-parametric families for ground- and space-based telescopes, with their advantages and limitations. Validation methods for PSF models are then addressed, with several metrics related to weak-lensing studies discussed in detail. Finally, we explore current challenges and future directions in PSF modelling for astronomical telescopes.</p>}}
+
+
+@ARTICLE{Wells.etal.1981,
+       author = {{Wells}, D.~C. and {Greisen}, E.~W. and {Harten}, R.~H.},
[…]
