### 2025-07-16 357b127
add Fruchter+2010 reference
