+++ paper/paper.bib
@@ -817,4 +817,16 @@ ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of par
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@INPROCEEDINGS{Fruchter.etal.2010,
+       author = {{Fruchter}, A.~S. and {et al.}},
+        title = "{BetaDrizzle: A Redesign of the MultiDrizzle Package}",
+     keywords = {betadrizzle, drizzlepac, astrodrizzle, drizzle, dither},
+    booktitle = {2010 Space Telescope Science Institute Calibration Workshop},
+         year = 2010,
+        month = jul,
+        pages = {382-387},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2010bdrz.conf..382F},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 
