### 2025-07-20 7ee6df1
hyperlink source repos per #12
