+++ paper/paper.bib
@@ -830,3 +830,32 @@ ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of par
 }
 
 
+@INPROCEEDINGS{Desjardins.etal.2023,
+       author = {{Desjardins}, Tyler and {Al-Kowsi}, Hanna and {Beaton}, Rachael and {Bellini}, Andrea and {Casertano}, Stefano and {Cosentino}, Richard and {Davis}, David and {Dencheva}, Nadezhda and {De Rosa}, Gisella and {Ferguson}, Henry and {Fleming}, Scott and {Gomez}, Sebastian and {Hargis}, Jonathan and {Hayden}, Brian and {Hoffmann}, Samantha and {Mutchler}, Maximilian and {Otor}, O. Justin and {Petric}, Andreea and {Sanchez}, Javier and {Schlafly}, Edward and {Sharma}, Sanjib and {Shen}, Sharon and {Snyder}, Gregory and {Sosey}, Megan and {Wu}, John and {the Roman Team}},
+        title = "{The Roman Space Telescope Science Operations Center: Wide Field Instrument Data Processing and Products}",
+    booktitle = {American Astronomical Society Meeting Abstracts},
+         year = 2023,
+       series = {American Astronomical Society Meeting Abstracts},
+       volume = {242},
+        month = jun,
+          eid = {230.01},
+        pages = {230.01},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023AAS...24223001D},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+@INPROCEEDINGS{Schultz.etal.2025,
+       author = {{Schultz}, William and {Otor}, Oderah and {Bellini}, Andrea and {Gomez}, Sebastian and {Han}, Eunkyu and {Lajoie}, Charles-Philippe and {Perrin}, Marshall and {Schlafly}, Edward and {the Roman Team}},
+        title = "{Desktop-based Simulation Tools for the Roman Space Telescope}",
+    booktitle = {American Astronomical Society Meeting Abstracts},
+         year = 2025,
+       series = {American Astronomical Society Meeting Abstracts},
+       volume = {246},
+        month = jun,
+          eid = {423.05},
+        pages = {423.05},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025AAS...24642305S},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
