+++ paper/paper.md
@@ -35,7 +35,7 @@ Given a directory containing reduced and calibrated, but not yet co-added, FITS 
 
 ![Comparison of drizzled PSFs generated for HST images using the default parameters for different methods included in `spike`. All panels use the same ACS/WFC imaging of the COSMOS field in F475W. The stretch of each model's colormap is chosen based on their individual handling of the background, wings, and faint features of the PSF. Note that the ePSF panel (second from right) shows some artifacts; the robustness of the effective PSF method is heavily dependent on the number of stars in the chosen field and may be changed by altering the star detection threshold. \label{fig:hstcomparison}](spike_psf_hstcompare.png)
 
-`spike` works with both model and empirical PSFs. The built-in PSF generation options are `TinyTim` [@Krist.etal.2011] and the @Gillis.etal.2020 modification, `STPSF` -- formerly `WebbPSF` [@Perrin_STPSF; @Perrin.etal.2012; @Perrin.etal.2014; @Perrin_WebbPSF], `photutils` ePSFs [@Anderson.King.2000; @Anderson.2016; @Bradley.etal], `PSFEx` [@Bertin.2011], and Space Telescope Science Institute's library of empirical STDPSFs for [HST](https://www.stsci.edu/~jayander/HST1PASS/LIB/PSFs/STDPSFs/) and [JWST](https://www.stsci.edu/~jayander/JWST1PASS/LIB/PSFs/STDPSFs/) [@Anderson.2016; @Libralato.etal.2023; @Libralato.etal.2024], all of which are included for having the ability to generate a model PSF for an arbitrary detector location. \autoref{fig:hstcomparison} compares output drizzled PSFs from the different methods for HST and \autoref{fig:jwstcomparison} compares output drizzled PSFs for JWST.
+`spike` works with both model and empirical PSFs. The built-in PSF generation options are Tiny Tim [@Krist.etal.2011] and the @Gillis.etal.2020 modification, STPSF -- formerly WebbPSF [@Perrin_STPSF; @Perrin.etal.2012; @Perrin.etal.2014; @Perrin_WebbPSF], Photutils ePSFs [@Anderson.King.2000; @Anderson.2016; @Bradley.etal], PSFEx [@Bertin.2011], and Space Telescope Science Institute's library of empirical STDPSFs for [HST](https://www.stsci.edu/~jayander/HST1PASS/LIB/PSFs/STDPSFs/) and [JWST](https://www.stsci.edu/~jayander/JWST1PASS/LIB/PSFs/STDPSFs/) [@Anderson.2016; @Libralato.etal.2023; @Libralato.etal.2024], all of which are included for having the ability to generate a model PSF for an arbitrary detector location. \autoref{fig:hstcomparison} compares output drizzled PSFs from the different methods for HST and \autoref{fig:jwstcomparison} compares output drizzled PSFs for JWST.
 
 ![Same as \autoref{fig:hstcomparison} for \textit{JWST}/NIRCam imaging in F115W. Note that the ePSF model shown here was generated using a lower detection threshold and a different star selection algorithm due to a paucity of stars in this field. \label{fig:jwstcomparison}](spike_psf_jwstcompare.png)
 
@@ -52,31 +52,31 @@ There is a preliminary module included in `spike` for recovering resampled Roman
[…]
