### 2025-07-20 14c48b9
remove repo only references
