+++ paper/paper.bib
@@ -455,32 +455,6 @@ year = {2024}
 }
 
 
-@misc{romancal,
-  author = {{Roman calibration pipeline developers}},
-  title = {romancal: {Python} library to process science observations from the {Nancy Grace Roman Space Telescope}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/romancal}
-}
-
-@misc{stcal,
-  author = {STScI},
-  title = {{STCAL: STScI Calibration algorithms and tools.}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/stcal}
-}
-
-@misc{roman_datamodels,
-  author = {STScI},
-  title = {roman\_datamodels: data models supporting calibration of the {Nancy Grace Roman Space Telescope}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/roman_datamodels}
-}
 
 @software{spherical_geometry,
   author       = {Bernie Simon and
@@ -521,32 +495,6 @@ year = {2024}
   url          = {https://doi.org/10.5281/zenodo.11636209}
 }
 
-@misc{stdatamodels,
-  author = {STScI},
-  title = {{stdatamodels}: Core support for {DataModel} classes used in calibration pipelines},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/stdatamodels}
-}
-
-@misc{psutil,
-  author = {Giampaolo Rodola},
-  title = {psutil: Cross-platform lib for process and system monitoring in {Python}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/giampaolo/psutil}
-}
-
-@misc{stpipe,
-  author = {STScI},
-  title = {{stpipe}: Framework for calibration pipeline software},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/spacetelescope/stpipe}
-}
 
 @ARTICLE{Greenfield.etal.2015,
        author = {{Greenfield}, P. and {Droettboom}, M. and {Bray}, E.},
@@ -622,14 +570,6 @@ year = {2024}
   url          = {https://doi.org/10.5281/zenodo.4018234}
 }
 
-@misc{pyyaml,
-  author = {Kirill Simonov},
-  title = {{PyYAML}: {YAML} parser and emitter for {Python}},
-  year = {2024},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/yaml/pyyaml}
-}
 
 @ARTICLE{Greenfield.Miller.2016,
        author = {{Greenfield}, P. and {Miller}, T.},
