### 2025-07-23 #8200 @ivastar
@avapolzin @warrickball

`stpsf` now has a Zenodo DOI: https://zenodo.org/records/15747364

also `roman_datamodels`: https://zenodo.org/records/16048709
