### 2025-07-23 e6dd21e
add new Zenodo records per openjournals/joss-reviews#8200
