+++ paper/paper.md
@@ -66,7 +66,7 @@ Since drizzled `spike` PSFs are intended for use with calibrated and resampled d
  - [PSFEx](https://github.com/astromatic/psfex?tab=readme-ov-file) [@Bertin.2011]
  - [psutil](https://github.com/giampaolo/psutil)
  - [PyYAML](https://github.com/yaml/pyyaml)
- - [`roman_datamodels`](https://github.com/spacetelescope/roman_datamodels)
+ - [`roman_datamodels`](https://github.com/spacetelescope/roman_datamodels) [@paul_huwe_2025_16048709]
  - [`romancal`](https://github.com/spacetelescope/romancal) [@Desjardins.etal.2023]
  - [SciPy](https://github.com/scipy/scipy) [@2020SciPy-NMeth]
  - [SExtractor](https://github.com/astromatic/sextractor) [@Bertin.Arnouts.1996]
@@ -76,7 +76,7 @@ Since drizzled `spike` PSFs are intended for use with calibrated and resampled d
  - [`stpipe`](https://github.com/spacetelescope/stpipe)
  - [Tiny Tim](https://github.com/spacetelescope/tinytim) [@Krist.etal.2011] -- including the option to use the @Gillis.etal.2020 parameters
  - [`tweakwcs`](https://github.com/spacetelescope/tweakwcs) [@cara.etal]
- - [STPSF](https://github.com/spacetelescope/stpsf) [@Schultz.etal.2025; @Perrin_STPSF], formerly [WebbPSF](https://github.com/spacetelescope/webbpsf) [@Perrin.etal.2012; @Perrin.etal.2014; @Perrin_WebbPSF]
+ - [STPSF](https://github.com/spacetelescope/stpsf) [@perrin_2025_15747364], formerly [WebbPSF](https://github.com/spacetelescope/webbpsf) [@Perrin.etal.2012; @Perrin.etal.2014; @Perrin_WebbPSF]
 
 # Acknowledgments
 
