+++ paper/paper.bib
@@ -715,14 +715,6 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@software{Perrin_STPSF,
-author = {Perrin, Marshall and Melendez, Marcio and Osborne, Shannon and Geda, Robel and Sappington, Bradley and Lajoie, Charles-Philippe and Long, Joseph and Otor, O. Justin},
-license = {BSD-3-Clause},
-title = {{STPSF}},
-url = {https://github.com/spacetelescope/stpsf},
-version = {2.0.0},
-year = 2025
-}
 
 @ARTICLE{Liaudat.etal.2023,
   
@@ -785,17 +777,68 @@ ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of par
 }
 
 
-@INPROCEEDINGS{Schultz.etal.2025,
-       author = {{Schultz}, William and {Otor}, Oderah and {Bellini}, Andrea and {Gomez}, Sebastian and {Han}, Eunkyu and {Lajoie}, Charles-Philippe and {Perrin}, Marshall and {Schlafly}, Edward and {the Roman Team}},
-        title = "{Desktop-based Simulation Tools for the Roman Space Telescope}",
-    booktitle = {American Astronomical Society Meeting Abstracts},
-         year = 2025,
-       series = {American Astronomical Society Meeting Abstracts},
-       volume = {246},
-        month = jun,
-          eid = {423.05},
-        pages = {423.05},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2025AAS...24642305S},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+
+@software{perrin_2025_15747364,
+  author       = {Perrin, Marshall and
+                  Long, Joseph and
+                  Osborne, Shannon and
+                  Geda, Robel and
+                  Sappington, Bradley and
+                  Meléndez, Marcio and
+                  Lajoie, Charles-Philippe and
+                  Leisenring, Jarron and
+                  Zimmerman, Neil and
+                  Brooks, Keira and
+                  Otor, O. Justin and
+                  Kulp, Trey and
+                  Chambers, Lauren and
+                  Jurling, Alden},
+  title        = {STPSF},
+  month        = jun,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {2.1.0},
+  doi          = {10.5281/zenodo.15747364},
+  url          = {https://doi.org/10.5281/zenodo.15747364},
+  swhid        = {swh:1:dir:5e5bd010e4d59bdba2972dbb87ac9974ded5c0b2
+                   ;origin=https://doi.org/10.5281/zenodo.15747363;vi
+                   sit=swh:1:snp:608253a7eb3b19a2393685b8cda0324adb08
+                   0b34;anchor=swh:1:rel:e6ec43de739ba24299dfd403e8ec
+                   3beeabd31119;path=spacetelescope-stpsf-a8a6e17
+                  },
 }
 
+@software{paul_huwe_2025_16048709,
+  author       = {Paul Huwe and
+                  William Jamieson and
+                  perrygreenfield and
+                  Zach Burnett and
+                  Brett Graham and
+                  Nadia Dencheva and
+                  Eddie Schlafly and
+                  Edward Slavich and
+                  D Davis and
+                  Jonathan Eisenhamer and
+                  Clare Shanahan and
+                  mairan and
+                  P. L. Lim and
+                  Larry Bradley and
+                  Mihai Cara and
+                  dmggh and
+                  Brett M. Morris and
+                  Zane Geiger},
[…]
