### 2025-07-24 #8200 @warrickball
Hi again @avapolzin. Sorry this is dragging out but note the errors above about missing references. These are the same things that show up as **?** in the [most recent PDF draft](https://github.com/openjournals/joss-reviews/issues/8200#issuecomment-3115220136), so that should help you debug the issue. You can re-run `@editorialbot generate pdf` (but not `recommend-accept`) but if you're uncomfortable with the "noise" of the extra comments, just let me know and I'll rerun the publication pipeline locally (though you'll have to wait until I get a chance).
