### 2025-07-24 2ca8584
swap out github -> zenodo references
