### 2025-07-24 #8200 @avapolzin
Thanks, @warrickball -- I'll rerun the `generate pdf` step. It seems I missed a couple of the GitHub -> new Zenodo DOI references.
