### 2025-07-24 1aa4d19
Spell out NASA ADS macro for FITS reference
