+++ paper/paper.bib
@@ -740,7 +740,7 @@ ABSTRACT={<p>The accurate modelling of the point spread function (PSF) is of par
 @ARTICLE{Wells.etal.1981,
        author = {{Wells}, D.~C. and {Greisen}, E.~W. and {Harten}, R.~H.},
         title = "{FITS - a Flexible Image Transport System}",
-      journal = {\aaps},
+      journal = {Astronomy and Astrophysics Supplement},
          year = 1981,
         month = jun,
        volume = {44},
