### 2025-07-24 5949187
protect Python in SciPy reference
