+++ paper/paper.bib
@@ -641,7 +641,7 @@ year = {2024}
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
   title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python},
+            Computing in {Python}},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
