### 2025-07-24 3426fff
add CITATION.cff
