+++ CITATION.cff
@@ -1,0 +1,32 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Polzin
+  given-names: Ava
+  orcid: "https://orcid.org/0000-0002-5283-933X"
+contact:
+- family-names: Polzin
+  given-names: Ava
+  orcid: "https://orcid.org/0000-0002-5283-933X"
+doi: 10.5281/zenodo.15791925
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Polzin
+    given-names: Ava
+    orcid: "https://orcid.org/0000-0002-5283-933X"
+  date-published: 2025-07-24
+  doi: 10.21105/joss.08200
+  issn: 2475-9066
+  issue: 111
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  start: 8200
+  title: "spike: A Tool to Drizzle HST, JWST, and Roman PSFs for
+    Improved Analyses"
+  type: article
+  url: "https://joss.theoj.org/papers/10.21105/joss.08200"
+  volume: 10
+title: "spike: A Tool to Drizzle HST, JWST, and Roman PSFs for Improved
+  Analyses"
