### 2025-07-25 #8200 @warrickball
Thanks once again to @EdwardBerman, @tobias-liaudat & @obi-wan76 for reviewing and @ivastar for editing this submission (and tireless editorial work across the journal)! JOSS simply wouldn't be possible without its community of volunteers.
